backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
httpx>=0.24.0
python-multipart>=0.0.5
gpxpy>=1.4.0
geopy>=2.3.0
python-dotenv>=0.19.0
